refac and implement simpleconsensus
