+++ src/eth/consensus/raft/tests/test_simple_blocks.rs
@@ -9,13 +9,13 @@ use tonic::Request;
 use super::factories::create_follower_consensus_with_leader;
 use super::factories::create_mock_block_entry;
 use super::factories::create_mock_transaction_execution_entry;
-use crate::eth::consensus::append_entry::append_entry_service_server::AppendEntryService;
-use crate::eth::consensus::append_entry::AppendBlockCommitRequest;
-use crate::eth::consensus::append_entry::AppendTransactionExecutionsRequest;
-use crate::eth::consensus::append_entry::StatusCode;
-use crate::eth::consensus::server::AppendEntryServiceImpl;
-use crate::eth::consensus::Role;
-use crate::eth::consensus::TransactionExecutionEntry;
+use crate::eth::consensus::raft::append_entry::append_entry_service_server::AppendEntryService;
+use crate::eth::consensus::raft::append_entry::AppendBlockCommitRequest;
+use crate::eth::consensus::raft::append_entry::AppendTransactionExecutionsRequest;
+use crate::eth::consensus::raft::append_entry::StatusCode;
+use crate::eth::consensus::raft::server::AppendEntryServiceImpl;
+use crate::eth::consensus::raft::Role;
+use crate::eth::consensus::raft::TransactionExecutionEntry;
 use crate::eth::primitives::Address;
 use crate::eth::primitives::BlockFilter;
 use crate::eth::primitives::BlockNumber;
