+++ src/infra/build_info.rs
@@ -1,7 +1,6 @@
 use serde_json::json;
 
 use crate::eth::rpc::RpcContext;
-use crate::eth::Consensus;
 use crate::ext::JsonValue;
 
 // -----------------------------------------------------------------------------
@@ -59,18 +58,6 @@ pub fn version() -> &'static str {
     }
 }
 
-pub fn is_leader() -> bool {
-    Consensus::is_leader()
-}
-
-pub fn current_term(ctx: &RpcContext) -> u64 {
-    ctx.consensus.current_term()
-}
-
-pub fn last_index(ctx: &RpcContext) -> u64 {
-    ctx.consensus.last_index()
-}
-
 /// Returns build info as JSON.
 pub fn as_json(ctx: &RpcContext) -> JsonValue {
     json!(
@@ -98,11 +85,6 @@ pub fn as_json(ctx: &RpcContext) -> JsonValue {
                 "version": RUST_VERSION,
                 "channel": RUST_CHANNEL,
                 "target": RUST_TARGET,
-            },
-            "consensus": {
-                "is_leader": is_leader(),
-                "current_term": current_term(ctx),
-                "last_index": last_index(ctx),
             }
         }
     )
